apps/api/requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.0.0
